+++ src/CameraHandling/ExtendedResponse.ts
@@ -1,0 +1,4 @@
+export interface ExtendedResponse {
+  responseTimestamp: number;
+  timeMillis: number;
+}