+++ src/CameraHandling/VideoConfig.ts
@@ -1,0 +1,20 @@
+export type VideoConfig = {
+  source?: string;
+  stillImageSource?: string;
+  returnAudioTarget?: string;
+  maxStreams?: number;
+  maxWidth?: number;
+  maxHeight?: number;
+  maxFPS?: number;
+  maxBitrate?: number;
+  forceMax?: boolean;
+  vcodec?: string;
+  packetSize?: number;
+  videoFilter?: string;
+  encoderOptions?: string;
+  mapvideo?: string;
+  mapaudio?: string;
+  audio?: boolean;
+  debug?: boolean;
+  debugReturn?: boolean;
+};